+++ index.html
@@ -7,7 +7,9 @@
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кекстаграм</title>
-  <script src="js/main.js"></script>
+  <script type="module" src="js/main.js"></script>
+  <script type="module" src="js/data.js"></script>
+  <script type="module" src="js/util.js"></script>
 </head>
 
 <body>
